refactor: reorganize validations evaluation folders for semantics
