+++ src/front-end/typescript/lib/pages/opportunity/sprint-with-us/edit/tab/overview.tsx
@@ -40,7 +40,7 @@ import {
 } from "shared/lib/resources/evaluations/sprint-with-us/team-questions";
 import { ADT, adt } from "shared/lib/types";
 import { isValid } from "shared/lib/validation";
-import { validateSWUTeamQuestionResponseEvaluationScores } from "shared/lib/validation/question-evaluation/sprint-with-us";
+import { validateSWUTeamQuestionResponseEvaluationScores } from "shared/lib/validation/evaluations/sprint-with-us/team-questions";
 
 export interface State extends Tab.Params {
   opportunity: SWUOpportunity | null;
