+++ src/front-end/typescript/lib/pages/proposal/sprint-with-us/view/tab/team-questions.tsx
@@ -49,7 +49,7 @@ import { invalid } from "shared/lib/validation";
 import {
   validateSWUTeamQuestionResponseEvaluationScoreNotes,
   validateSWUTeamQuestionResponseEvaluationScoreScore
-} from "shared/lib/validation/question-evaluation/sprint-with-us";
+} from "shared/lib/validation/evaluations/sprint-with-us/team-questions";
 import { makeStartLoading, makeStopLoading } from "front-end/lib";
 import { leftPlacement, iconLinkSymbol } from "front-end/lib/views/link";
 import Icon from "front-end/lib/views/icon";
